Move library code into package
