+++ python/aux/extract_weights.py
