+++ python/nemo/images.py
